+++ src/conservative_nn/lagrangian_dynamical_system.py
@@ -1,7 +1,7 @@
 """Classes for deriving equations of motion from a Lagrangian via automatic differentiation
 
 The constructor accepts a lagrangian object, which can for example be an instance of
-one of the classes in nn_models.py. It then uses automatic differentiation to
+one of the classes in nn_lagrangian_model.py. It then uses automatic differentiation to
 obtain the acceleration as a function of the position and velocity.
 
 The class LagrangianDynamicalSystem represent a generic dynamical system and assumes that
