+++ src/conservative_nn/nn_lagrangian_model.py
